frontEnd-Streamlit/requirements.txt
streamlit>=1.35,<2.0
pandas>=2.2,<3.0
plotly>=5.22,<6.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
